+++ src/main/java/baseball/exception/DuplicationNumberException.java
@@ -1,4 +1,4 @@
-package baseball.domain.exception;
+package baseball.exception;
 
 public class DuplicationNumberException extends RuntimeException {
