+++ src/main/java/baseball/exception/NumberRangeInvalidException.java
@@ -1,4 +1,4 @@
-package baseball.domain.exception;
+package baseball.exception;
 
 public class NumberRangeInvalidException extends RuntimeException {
